Move path::Id to core
